Improve auto-backup efficiency and only run on Leader

+++ auto/backup/uploader.go
@@ -23,7 +23,14 @@ type StorageClient interface {
 // service will call Provide() to have the data-for-upload to be written to the
 // to the file specified by path.
 type DataProvider interface {
-	Provide(path string) error
+	// LastModified returns the time the data managed by the DataProvider was
+	// last modified.
+	LastModified() (time.Time, error)
+
+	// Provide writes the data-for-upload to the file specified by path. Because
+	// Provide may change the data in the DataProvider, it returns the current
+	// modified time of the data, after the data has been written to path.
+	Provide(path string) (time.Time, error)
 }
 
 // stats captures stats for the Uploader service.
@@ -66,11 +73,7 @@ type Uploader struct {
 	lastUploadTime     time.Time
 	lastUploadDuration time.Duration
 
-	lastSum SHA256Sum
-
-	// disableSumCheck is used for testing purposes to disable the check that
-	// prevents uploading the same data twice.
-	disableSumCheck bool
+	lastModified time.Time
 }
 
 // NewUploader creates a new Uploader service.
@@ -103,10 +106,6 @@ func (u *Uploader) Start(ctx context.Context, isUploadEnabled func() bool) chan 
 				return
 			case <-ticker.C:
 				if !isUploadEnabled() {
-					// Reset the lastSum so that the next time we're enabled upload will
-					// happen. We do this to be conservative, as we don't know what was
-					// happening while upload was disabled.
-					u.lastSum = nil
 					continue
 				}
 				if err := u.upload(ctx); err != nil {
@@ -126,33 +125,38 @@ func (u *Uploader) Stats() (map[string]interface{}, error) {
 		"compress":             u.compress,
 		"last_upload_time":     u.lastUploadTime.Format(time.RFC3339),
 		"last_upload_duration": u.lastUploadDuration.String(),
-		"last_upload_sum":      u.lastSum.String(),
+		"last_modified":        u.lastModified.String(),
 	}
 	return status, nil
 }
 
 func (u *Uploader) upload(ctx context.Context) error {
-	// create a temporary file for the data to be uploaded
+	var err error
+	var lm time.Time
+
+	// Data source changed?
+	lm, err = u.dataProvider.LastModified()
+	if err != nil {
+		return err
+	}
+	if !lm.After(u.lastModified) {
+		stats.Add(numUploadsSkipped, 1)
+		return nil
+	}
+
+	// Create a temporary file for the data to be uploaded
 	filetoUpload, err := tempFilename()
 	if err != nil {
 		return err
 	}
 	defer os.Remove(filetoUpload)
 
-	if err := u.dataProvider.Provide(filetoUpload); err != nil {
+	lm, err = u.dataProvider.Provide(filetoUpload)
+	if err != nil {
 		return err
 	}
 	if err := u.compressIfNeeded(filetoUpload); err != nil {
 		return err
-	}
-
-	sum, err := FileSHA256(filetoUpload)
-	if err != nil {
-		return err
-	}
-	if !u.disableSumCheck && sum.Equals(u.lastSum) {
-		stats.Add(numUploadsSkipped, 1)
-		return nil
 	}
 
 	fd, err := os.Open(filetoUpload)
@@ -167,7 +171,7 @@ func (u *Uploader) upload(ctx context.Context) error {
 	if err != nil {
 		stats.Add(numUploadsFail, 1)
 	} else {
-		u.lastSum = sum
+		u.lastModified = lm
 		stats.Add(numUploadsOK, 1)
 		stats.Add(totalUploadBytes, cr.Count())
 		stats.Get(lastUploadBytes).(*expvar.Int).Set(cr.Count())
